requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
defusedxml==0.7.1
Django==4.2.7
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-cuser==2017.3.16
django-filter==25.1
django-jazzmin==2.6.0
django-js-asset==3.0.1
django-modeltranslation==0.19.12
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
joblib==1.4.2
modeltranslation==0.25
nltk==3.9.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
